# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CourierPrimaryKey:
    properties:
      id:
        type: string
    type: object
...
  models.CreateCourier:
    properties:
      name:
        type: string
      phone:
        type: string
    type: object
...
  models.PatchRequest:
    properties:
      fields:
        additionalProperties: true
        type: object
      id:
        type: string
    type: object
...
  models.UpdateCourier:
    properties:
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /courier:
    get:
      consumes:
      - application/json
      description: Get List Courier
      operationId: get_list_courier
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Courier
      tags:
      - Courier
    post:
      consumes:
      - application/json
      description: Create Courier
      operationId: create_courier
      parameters:
      - description: CreateCourierRequest
        in: body
        name: Courier
        required: true
        schema:
          $ref: '#/definitions/models.CreateCourier'
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Courier
      tags:
      - Courier
  /courier/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Courier
      operationId: delete_courier
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: DeleteCourierRequest
        in: body
        name: courier
        required: true
        schema:
          $ref: '#/definitions/models.CourierPrimaryKey'
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Courier
      tags:
      - Courier
    get:
      consumes:
      - application/json
      description: Get By ID Courier
      operationId: get_by_id_Courier
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID Courier
      tags:
      - Courier
    patch:
      consumes:
      - application/json
      description: Update Patch Courier
      operationId: updat_patch_courier
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdatePatchCourierRequest
        in: body
        name: courier
        required: true
        schema:
          $ref: '#/definitions/models.PatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Patch Courier
      tags:
      - Courier
    put:
      consumes:
      - application/json
      description: Update Courier
      operationId: update_courier
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateCourierRequest
        in: body
        name: courier
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCourier'
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Courier
      tags:
      - Courier
...
swagger: "2.0"